Refactor ProjectWorkspace to multiple widgets
