+++ hailstorm-web-client/e2e/features/support/pages/ClusterConfig.ts
@@ -2,6 +2,11 @@ import { WizardBase } from './WizardBase';
 
 export class ClusterConfig extends WizardBase {
 
+  proceedToNextStep() {
+    browser.waitUntil(() => this.nextButton.isEnabled(), 15000);
+    this.nextButton.click();
+  }
+
   protected chooseCluster(clusterLink: WebdriverIO.Element) {
     browser.waitUntil(() => browser.react$("Loader").isDisplayed() === false, 10000);
     if (clusterLink.isDisplayed()) {
