+++ hailstorm-web-client/e2e/features/support/pages/ProjectWorkspace.ts
@@ -7,12 +7,6 @@ export class ProjectWorkspace {
   get checkBoxes() { return $$('//tbody//tr/td/input'); }
   get jMeterPlansEdit() { return $('//*[@data-testid="JMeter Plans"]//button'); }
   get abortButton() { return $('button*=Abort'); }
-  get showDangerousSettings() { return $('button*=Show them'); }
-  get terminateButton() { return $('button*=Terminate this project'); }
-  get confirmTerminate() { return $('button*=Yes, Terminate'); }
-  get masterCheckBox() { return $('//table/thead/tr/th/input'); }
-  get reportButton() { return $('button*=Report'); }
-  get reportsListItems() { return $$('//*[@data-testid="Reports List"]/a'); }
 
   startTest() {
     browser.waitUntil(() => this.startButton.isEnabled(), 15000);
@@ -62,30 +56,6 @@ export class ProjectWorkspace {
     setTimeout(() => {
       abortButton.click();
     }, seconds * 1000);
-  }
-
-  terminateProject() {
-    this.showDangerousSettings.click();
-    this.terminateButton.waitForDisplayed(1000);
-    this.terminateButton.click();
-    this.confirmTerminate.waitForDisplayed(1000);
-    this.confirmTerminate.click();
-    if (!this.terminateButton.isEnabled()) {
-      this.terminateButton.waitForEnabled(1000, true);
-      browser.waitUntil(() => this.terminateButton.isEnabled(), 5 * 60 * 1000, "wait for terminate action to complete", 3000);
-    }
-  }
-
-  generateReport() {
-    browser.pause(500);
-    this.masterCheckBox.click();
-    this.reportButton.waitForEnabled(5000);
-    this.reportButton.click();
-  }
-
-  waitForGeneratedReports() {
-    browser.waitUntil(() => this.reportsListItems.length > 0, 5 * 60 * 1000, "waiting for report to be generated", 10000);
-    return this.reportsListItems.length;
   }
 
   projectIdFromUrl(): string {
